+++ spec/support/flores_extensions.rb
@@ -1,49 +1,68 @@
 # encoding: utf-8
 require "flores/pki"
 require "flores/random"
+require "socket"
 
 module Flores
   module Random
     DEFAULT_PORT_RANGE = 1024..65535
-    DEFAULT_PORT_CHECK_TIMEOUT = 1
-    DEFAULT_MAXIMUM_PORT_FIND_TRY = 15
-
     class << self
       def port(range = DEFAULT_PORT_RANGE)
-        try = 0
-        while try < DEFAULT_MAXIMUM_PORT_FIND_TRY
-          candidate = integer(range)
-
-          if port_available?(candidate)
-            break
-          else
-            try += 1
-          end
-        end
-        
-        raise "Flores.random_port: Cannot find an available port, tried #{DEFAULT_MAXIMUM_PORT_FIND_TRY} times, range was: #{range}" if try == DEFAULT_MAXIMUM_PORT_FIND_TRY
-
-        candidate
-      end
-      
-      def port_available?(port)
-        begin
-          server = TCPServer.new(port)
-          available = true
-        rescue # Assume that any errors can do this
-          available = false
-        ensure
-          server.close if server
-        end
-
-        return available
+        integer(range)
       end
     end
   end
 
   module PKI
+
+    # Monkey patched the fix for https://github.com/jordansissel/ruby-flores/issues/9
+    # TODO: remove this once Flores is released with fix.
+    CertificateSigningRequest.class_eval do
+      def create
+        validate!
+        extensions = OpenSSL::X509::ExtensionFactory.new
+        extensions.subject_certificate = certificate
+        extensions.issuer_certificate = self_signed? ? certificate : signing_certificate
+
+        certificate.issuer = extensions.issuer_certificate.subject
+        certificate.add_extension(extensions.create_extension("subjectKeyIdentifier", "hash", false))
+
+        if want_signature_ability?
+          # Create a CA.
+          certificate.add_extension(extensions.create_extension("basicConstraints", "CA:TRUE", true))
+          # Rough googling seems to indicate at least keyCertSign is required for CA and intermediate certs.
+          certificate.add_extension(extensions.create_extension("keyUsage", "keyCertSign, cRLSign, digitalSignature", true))
+        else
+          # Create a client+server certificate
+          #
+          # It feels weird to create a certificate that's valid as both server and client, but a brief inspection of major
+          # web properties (apple.com, google.com, yahoo.com, github.com, fastly.com, mozilla.com, amazon.com) reveals that
+          # major web properties have certificates with both clientAuth and serverAuth extended key usages. Further,
+          # these major server certificates all have digitalSignature and keyEncipherment for key usage.
+          #
+          # Here's the command I used to check this:
+          #    echo mozilla.com apple.com github.com google.com yahoo.com fastly.com elastic.co amazon.com \
+          #    | xargs -n1 sh -c 'openssl s_client -connect $1:443 \
+          #    | sed -ne "/-----BEGIN CERTIFICATE-----/,/-----END CERTIFICATE-----/p" \
+          #    | openssl x509 -text -noout | sed -ne "/X509v3 extensions/,/Signature Algorithm/p" | sed -e "s/^/$1 /"' - \
+          #    | grep -A2 'Key Usage'
+          certificate.add_extension(extensions.create_extension("keyUsage", "digitalSignature, keyEncipherment", true))
+          certificate.add_extension(extensions.create_extension("extendedKeyUsage", "clientAuth, serverAuth", false))
+        end
+
+        if @subject_alternates
+          certificate.add_extension(extensions.create_extension("subjectAltName", @subject_alternates.join(",")))
+        end
+
+        certificate.serial = OpenSSL::BN.new(serial)
+        certificate.sign(signing_key, digest_method)
+        certificate
+      end
+    end
+
+
     DEFAULT_CERTIFICATE_OPTIONS = {
-      :duration => Flores::Random.number(100..2000),
+      :duration => 86400, #one day
       :key_size => GENERATE_DEFAULT_KEY_SIZE, 
       :exponent => GENERATE_DEFAULT_EXPONENT,
       :want_signature_ability => false
@@ -78,5 +97,9 @@ module Flores
 
       [csr.create, client_key]
     end
+
+
+
   end
 end
+