+++ CHANGELOG.md
@@ -1,11 +1,29 @@
+## 4.0.3
+ - Include remote ip_address in metadata. #180
+ - Require Java 8 #221
+ - Fix ability to set SSL protocols  #228
+
+## 4.0.2
+  - Relax version of concurrent-ruby to `~> 1.0` #216
+
 ## 4.0.1
- - Breaking change: Logstash will no longer start when multiline codec is used with the Beats input plugin
+ - Breaking change: Logstash will no longer start when multiline codec is used with the Beats input plugin #201
 
 ## 4.0.0
- - Verion yanked from RubyGems for packaging issues
+ - Version yanked from RubyGems for packaging issues
+
+## 3.1.19 
+   - Fix ability to set SSL protocols  #228
+   
+## 3.1.18
+  - Relax version of concurrent-ruby to ~> 1.0 #216
+  
+## 3.1.17 
+ - Docs: Add note indicating that the multiline codec should not be used with the Beats input plugin
+ - Deprecate warning for multiline codec with the Beats input plugin
 
 ## 3.1.16
- - Docs: Add note indicating that the multiline codec should not be used with the Beats input plugin
+ - Version yanked from RubyGems for packaging issues
 
 ## 3.1.15
  - DEBUG: Add information about the remote when an exception is catched #192